Use texelFetch instead of texture2D in shaders requiring atlases

+++ client/shaders/block_hw/opengl_vertex.glsl
@@ -1,7 +1,7 @@
 layout (location = 0) in vec3 pos;
 layout (location = 1) in vec4 color;
 layout (location = 2) in vec3 normal;
-layout (location = 3) in vec2 uv;
+layout (location = 3) in ivec2 uv;
 layout (location = 4) in int materialType;
 layout (location = 5) in vec3 hwcolor;
 
@@ -31,9 +31,9 @@ out lowp vec3 vHWColor;
 // lie within the same bounds when MSAA is en- and disabled.
 // This fixes the stripes problem with nearest-neighbor textures and MSAA.
 #ifdef GL_ES
-out mediump vec2 vTexCoord;
+out mediump ivec2 vTexCoord;
 #else
-centroid out vec2 vTexCoord;
+centroid out ivec2 vTexCoord;
 #endif
 #ifdef ENABLE_DYNAMIC_SHADOWS
 	out float vCosLight;
